modifica finale codice funzionante!

+++ Progect_GreenEconomy_4BII/EnviromentData.cs
@@ -8,13 +8,13 @@ namespace Progect_GreenEconomy_4BII
     {
         public DateTime DataOra { get; set; }
         public double temperatura { get; set; }
-        public int qualitaAria { get; set; } // 1-5
+        public double qualitaAria { get; set; }
 
-        
+
         public EnviromentData() { } // costruttore vuoto per la serializzazione JSON
 
 
-        public EnviromentData(double temp, int aqi)
+        public EnviromentData(double temp, double aqi)
         {
             this.DataOra = DateTime.Now;
             this.temperatura = temp;
